requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain-openai
langchain-community
langchain-text-splitters
langchain-chroma
pypdf
